requirements.txt
python-telegram-bot
python-dotenv
beautifulsoup4
requests